# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/events:
    get:
      tags:
        - Events
      summary: List Events
      description: Get a paginated list of Events with optional filtering and sorting
      operationId: listEvents
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Events
      summary: Create Event
      description: Create a new Event
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/event_types:
    get:
      tags:
        - EventTypes
      summary: List EventTypes
      description: Get a paginated list of EventTypes with optional filtering and sorting
      operationId: listEventTypes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventTypeList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - EventTypes
      summary: Create EventType
      description: Create a new EventType
      operationId: createEventType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventTypeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Event:
      type: object
      required:
        - id
        - name
        - stage_id
        - kanban_state
        - date_begin
        - date_end
        - date_tz
        - is_multi_slots
        - event_slot_count
        - seats_limited
        - seats_max
        - badge_format
        - is_published
        - metadata
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        event_type_id:
          nullable: true
          type: string
          format: uuid
        stage_id:
          type: string
          format: uuid
        kanban_state:
          type: string
        date_begin:
          type: string
          format: date-time
        date_end:
          type: string
          format: date-time
        date_tz:
          type: string
          maxLength: 60
        is_multi_slots:
          type: boolean
        event_slot_count:
          type: integer
          format: int32
        seats_limited:
          type: boolean
        seats_max:
          type: integer
          format: int32
        address_id:
          nullable: true
          type: string
          format: uuid
        organizer_id:
          nullable: true
          type: string
          format: uuid
        user_id:
          nullable: true
          type: string
          format: uuid
        company_id:
          nullable: true
          type: string
          format: uuid
        event_url:
          nullable: true
          type: string
          format: uri
        badge_format:
          type: string
        is_published:
          type: boolean
        date_publish:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateEventRequest:
      type: object
      required:
        - name
        - stage_id
        - kanban_state
        - date_begin
        - date_end
        - date_tz
        - is_multi_slots
        - event_slot_count
        - seats_limited
        - seats_max
        - badge_format
        - is_published
        - metadata
      properties:
        name:
          type: string
          maxLength: 200
        event_type_id:
          nullable: true
          type: string
          format: uuid
        stage_id:
          type: string
          format: uuid
        kanban_state:
          type: string
        date_begin:
          type: string
          format: date-time
        date_end:
          type: string
          format: date-time
        date_tz:
          type: string
          maxLength: 60
        is_multi_slots:
          type: boolean
        event_slot_count:
          type: integer
          format: int32
        seats_limited:
          type: boolean
        seats_max:
          type: integer
          format: int32
        address_id:
          nullable: true
          type: string
          format: uuid
        organizer_id:
          nullable: true
          type: string
          format: uuid
        user_id:
          nullable: true
          type: string
          format: uuid
        company_id:
          nullable: true
          type: string
          format: uuid
        event_url:
          nullable: true
          type: string
          format: uri
        badge_format:
          type: string
        is_published:
          type: boolean
        date_publish:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    EventList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    EventType:
      type: object
      required:
        - id
        - name
        - seats_limited
        - seats_max
        - default_date_tz
        - is_multi_slots
        - metadata
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        description:
          nullable: true
          type: string
          maxLength: 2000
        seats_limited:
          type: boolean
        seats_max:
          type: integer
          format: int32
        default_date_tz:
          type: string
          maxLength: 60
        is_multi_slots:
          type: boolean
        metadata:
          type: object
...
    CreateEventTypeRequest:
      type: object
      required:
        - name
        - seats_limited
        - seats_max
        - default_date_tz
        - is_multi_slots
        - metadata
      properties:
        name:
          type: string
          maxLength: 120
        description:
          nullable: true
          type: string
          maxLength: 2000
        seats_limited:
          type: boolean
        seats_max:
          type: integer
          format: int32
        default_date_tz:
          type: string
          maxLength: 60
        is_multi_slots:
          type: boolean
        metadata:
          type: object
...
    EventTypeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EventType'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []